hang subsystem
the logic is a bit weird, but I think it should work. also I way over-commented so delete once works.

HangSubsystemTeleOp:

I wasn't sure what to do because this is an automated sequence and not incremental like the other ones so I kinda just guessed...

and I commented everything just incase.